Fix Sonar issues
There is no requirement that class names be unique, only that they be unique within a package. Therefore trying to determine an object's type based on its class name is an exercise fraught with danger. One of those dangers is that a malicious user will send objects of the same name as the trusted class and thereby gain trusted access.
Instead, the instanceof operator or the Class.isAssignableFrom() method should be used to check the object's underlying type.
Issue-ID: SDC-2697

Signed-off-by: shrek2000 <oren.kleks@amdocs.com>
Change-Id: Ie4fd94618135b425a7e505992649f1a6384b0f98