Fix tests not running after junit5 upgrade
There was a conflict with the junit version in spring boot and
the one in the project. With that, maven surefire plugin was not
finding the tests.
The org.onap.sdc.security.logging.wrappers.LoggerTest was also broken
with a conflict with mockito and an undesirable powermock library from
org.onap.portal.sdk:epsdk-fw artifact. There was also an incorrect
assertion in the test.

Issue-ID: SDC-3604
Signed-off-by: André Schmid <andre.schmid@est.tech>
Change-Id: Ieacedaa9c5204c2eab6ee96870f9b7726e06fc43